// src/lib/sdk/models/CreateCategoryDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreateCategoryDto = {
    /**
     * Merchant ID that owns this category
     */
    merchantId: string;
    /**
     * Name of the category
     */
    name: string;
    /**
     * Category description
     */
    description?: string;
    /**
     * Whether this category is active
     */
    isActive: boolean;
};