// src/lib/sdk/models/UpdateCategoryDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UpdateCategoryDto = {
    /**
     * Merchant ID that owns this category
     */
    merchantId?: string;
    /**
     * Name of the category
     */
    name?: string;
    /**
     * Category description
     */
    description?: string;
    /**
     * Whether this category is active
     */
    isActive?: boolean;
};